+++ Assets/Prefabs/SearchWorld/Legacy/1stStage.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 02d0b67940e11440ba3f33897dbfc4f9
+guid: 957e9e604809e44da82b9a92363ad95c
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
