+++ Assets/Prefabs/SearchWorld/Legacy/2ndStage.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 69b9e6198c78c4fd881c1087d3223f3a
+guid: 2592ac4fb77994b3f9050f6caab7c8f3
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
