+++ Assets/Scripts/GameLogic/Legacy_SearchWorld.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 69b9e6198c78c4fd881c1087d3223f3a
+guid: 0cf99ce0868dc4a898a92c06812d8dd4
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
