+++ Assets/Scripts/GameLogic/Legacy_SearchWorld/Effects/ZoomEffect.cs
