+++ Assets/Scripts/GameLogic/Legacy_SearchWorld/SearchWorldManager.cs
@@ -1,0 +1,14 @@
+using TMPro;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class SearchWorldManager : MonoBehaviour
+{
+    [SerializeField] private SearchWorld searchWorld;
+    [SerializeField] private TextMeshProUGUI worldName;
+
+    void Start()
+    {
+        worldName.text = searchWorld.WorldName;
+    }
+}
