+++ Assets/Scripts/GameLogic/Legacy_SearchWorld/ZoomedView.cs
