--- Assets/Scripts/GameLogic/SearchWorld/SearchWorldManager.cs
@@ -1,14 +1,0 @@
-using TMPro;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class SearchWorldManager : MonoBehaviour
-{
-    [SerializeField] private SearchWorld searchWorld;
-    [SerializeField] private TextMeshProUGUI worldName;
-
-    void Start()
-    {
-        worldName.text = searchWorld.WorldName;
-    }
-}
